Migrated Kafka SSL ++ Grateful Shutdown
++ Usage of `@DisabledOnNative` (from test framework) instead of `@DisabledOnNativeImage` (from Quarkus upstream)
++ Rename packages to remove the `.openshift.` part.
