+++ messaging/kafka-avro-reactive-messaging/src/main/java/io/quarkus/ts/messaging/kafka/StockPriceEndpoint.java
@@ -1,0 +1,27 @@
+package io.quarkus.ts.messaging.kafka;
+
+import javax.inject.Inject;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+
+import org.eclipse.microprofile.reactive.messaging.Channel;
+import org.jboss.resteasy.annotations.SseElementType;
+import org.reactivestreams.Publisher;
+
+@Path("/stock-price")
+public class StockPriceEndpoint {
+
+    @Inject
+    @Channel("price-stream")
+    Publisher<String> stockPrices;
+
+    @GET
+    @Path("/stream")
+    @Produces(MediaType.SERVER_SENT_EVENTS)
+    @SseElementType(MediaType.APPLICATION_JSON)
+    public Publisher<String> stream() {
+        return stockPrices;
+    }
+}
