+++ messaging/kafka-avro-reactive-messaging/src/test/java/io/quarkus/ts/messaging/kafka/BaseKafkaAvroIT.java
@@ -1,0 +1,65 @@
+package io.quarkus.ts.messaging.kafka;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+
+import javax.ws.rs.client.Client;
+import javax.ws.rs.client.ClientBuilder;
+import javax.ws.rs.client.WebTarget;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.sse.SseEventSource;
+
+import org.junit.jupiter.api.Test;
+
+import io.quarkus.test.bootstrap.RestService;
+
+public abstract class BaseKafkaAvroIT {
+
+    private static final int TIMEOUT_SEC = 25;
+    private static final int EVENTS_AMOUNT = 3;
+
+    private String endpoint;
+    private Client client = ClientBuilder.newClient();
+    private List<String> receive = new CopyOnWriteArrayList<>();
+    private boolean completed;
+
+    @Test
+    public void testAlertMonitorEventStream() throws InterruptedException {
+        givenAnApplicationEndpoint(getEndpoint() + "/stock-price/stream");
+        whenRequestSomeEvents(EVENTS_AMOUNT);
+        thenVerifyAllEventsArrived();
+    }
+
+    protected abstract RestService getApp();
+
+    private void givenAnApplicationEndpoint(String endpoint) {
+        this.endpoint = endpoint;
+    }
+
+    private void whenRequestSomeEvents(int amount) throws InterruptedException {
+        WebTarget target = client.target(endpoint);
+        CountDownLatch latch = new CountDownLatch(amount);
+
+        SseEventSource source = SseEventSource.target(target).build();
+        source.register(inboundSseEvent -> {
+            receive.add(inboundSseEvent.readData(String.class, MediaType.APPLICATION_JSON_TYPE));
+            latch.countDown();
+        });
+
+        source.open();
+        completed = latch.await(TIMEOUT_SEC, TimeUnit.SECONDS);
+        source.close();
+    }
+
+    private void thenVerifyAllEventsArrived() {
+        assertTrue(completed, "Not all expected kafka events has been consumed.");
+    }
+
+    private String getEndpoint() {
+        return getApp().getHost() + ":" + getApp().getPort();
+    }
+}
