+++ messaging/kafka-avro-reactive-messaging/src/test/java/io/quarkus/ts/messaging/kafka/OpenShiftAmqStreamsKafkaAvroIT.java
@@ -1,0 +1,25 @@
+package io.quarkus.ts.messaging.kafka;
+
+import io.quarkus.test.bootstrap.KafkaService;
+import io.quarkus.test.bootstrap.RestService;
+import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.services.KafkaContainer;
+import io.quarkus.test.services.QuarkusApplication;
+import io.quarkus.test.services.containers.model.KafkaVendor;
+
+@OpenShiftScenario
+public class OpenShiftAmqStreamsKafkaAvroIT extends BaseKafkaAvroIT {
+
+    @KafkaContainer(vendor = KafkaVendor.STRIMZI, image = "${amq-streams.image}", version = "${amq-streams.version}", withRegistry = true)
+    static KafkaService kafka = new KafkaService();
+
+    @QuarkusApplication
+    static RestService app = new RestService()
+            .withProperty("kafka.bootstrap.servers", kafka::getBootstrapUrl)
+            .withProperty("kafka.registry.url", kafka::getRegistryUrl);
+
+    @Override
+    public RestService getApp() {
+        return app;
+    }
+}
