+++ messaging/kafka-avro-reactive-messaging/src/test/java/io/quarkus/ts/messaging/kafka/OpenShiftStrimziKafkaAvroIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.ts.messaging.kafka;
+
+import io.quarkus.test.scenarios.OpenShiftScenario;
+
+@OpenShiftScenario
+public class OpenShiftStrimziKafkaAvroIT extends StrimziKafkaAvroIT {
+}
