--- messaging/kafka-avro-reactive-messaging/src/test/java/io/quarkus/ts/openshift/messaging/kafka/OpenShiftStrimziKafkaAvroIT.java
@@ -1,7 +1,0 @@
-package io.quarkus.ts.openshift.messaging.kafka;
-
-import io.quarkus.test.scenarios.OpenShiftScenario;
-
-@OpenShiftScenario
-public class OpenShiftStrimziKafkaAvroIT extends StrimziKafkaAvroIT {
-}
