--- messaging/kafka-avro-reactive-messaging/src/test/java/io/quarkus/ts/openshift/messaging/kafka/StrimziKafkaAvroIT.java
@@ -1,26 +1,0 @@
-package io.quarkus.ts.openshift.messaging.kafka;
-
-import io.quarkus.test.bootstrap.KafkaService;
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.KafkaContainer;
-import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.test.services.containers.model.KafkaVendor;
-
-@QuarkusScenario
-public class StrimziKafkaAvroIT extends BaseKafkaAvroIT {
-
-    @KafkaContainer(vendor = KafkaVendor.STRIMZI, withRegistry = true)
-    static KafkaService kafka = new KafkaService();
-
-    @QuarkusApplication
-    static RestService app = new RestService()
-            .withProperties("strimzi-application.properties")
-            .withProperty("kafka.bootstrap.servers", kafka::getBootstrapUrl)
-            .withProperty("kafka.registry.url", kafka::getRegistryUrl);
-
-    @Override
-    public RestService getApp() {
-        return app;
-    }
-}
