+++ messaging/kafka-streams-reactive-messaging/src/main/resources/application.properties
@@ -1,18 +1,30 @@
 producer.httpCodes=200,201,400,401,403,500
 producer.loginUrls=redhat/login,inditex/login,santander/login,bbva/login
+login.denied.windows.sec=3
 # TODO: Disable health check because of https://github.com/quarkusio/quarkus/issues/15464
 quarkus.reactive-messaging.health.enabled=false
+
 mp.messaging.outgoing.login-http-response-values.connector=smallrye-kafka
 mp.messaging.outgoing.login-http-response-values.key.serializer=org.apache.kafka.common.serialization.StringSerializer
 mp.messaging.outgoing.login-http-response-values.value.serializer=org.apache.kafka.common.serialization.StringSerializer
+
 mp.messaging.incoming.login-denied.connector=smallrye-kafka
 mp.messaging.incoming.login-denied.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
 mp.messaging.incoming.login-denied.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
 mp.messaging.incoming.login-denied.enable.auto.commit=true
+
+# Kafka shutdown config
+mp.messaging.outgoing.slow-topic.connector=smallrye-kafka
+mp.messaging.outgoing.slow-topic.topic=slow
+mp.messaging.outgoing.slow-topic.value.serializer=org.apache.kafka.common.serialization.StringSerializer
+
+mp.messaging.incoming.slow.connector=smallrye-kafka
+mp.messaging.incoming.slow.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+
 quarkus.kafka-streams.application-id=login-denied-aggregator
 quarkus.kafka-streams.application-server=localhost:${quarkus.http.port}
 quarkus.kafka-streams.topics=login-http-response-values
-login.denied.windows.sec=3
+
 # streams options
 kafka-streams.cache.max.bytes.buffering=10240
 kafka-streams.commit.interval.ms=1000
@@ -20,3 +32,9 @@ kafka-streams.metadata.max.age.ms=500
 kafka-streams.auto.offset.reset=earliest
 kafka-streams.consumer.session.timeout.ms=10000
 kafka-streams.consumer.heartbeat.interval.ms=80
+
+# Native
+quarkus.native.enable-all-security-services=true
+
+# TODO: Workaround for https://github.com/quarkusio/quarkus/issues/19296
+quarkus.kafka.devservices.enabled=false
