+++ messaging/kafka-streams-reactive-messaging/src/test/java/io/quarkus/ts/messaging/kafka/StrimziKafkaStreamIT.java
@@ -1,0 +1,25 @@
+package io.quarkus.ts.messaging.kafka;
+
+import io.quarkus.test.bootstrap.KafkaService;
+import io.quarkus.test.bootstrap.RestService;
+import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.services.KafkaContainer;
+import io.quarkus.test.services.QuarkusApplication;
+import io.quarkus.test.services.containers.model.KafkaVendor;
+
+@QuarkusScenario
+public class StrimziKafkaStreamIT extends BaseKafkaStreamTest {
+
+    @KafkaContainer(vendor = KafkaVendor.STRIMZI)
+    static KafkaService kafka = new KafkaService();
+
+    @QuarkusApplication
+    static RestService app = new RestService()
+            .withProperty("kafka.bootstrap.servers", kafka::getBootstrapUrl)
+            .withProperty("quarkus.kafka-streams.bootstrap-servers", kafka::getBootstrapUrl);
+
+    @Override
+    protected String getAppUrl() {
+        return app.getHost() + ":" + app.getPort();
+    }
+}
