--- messaging/kafka-streams-reactive-messaging/src/test/java/io/quarkus/ts/openshift/messaging/kafka/OpenShiftStrimziKafkaStreamIT.java
@@ -1,7 +1,0 @@
-package io.quarkus.ts.openshift.messaging.kafka;
-
-import io.quarkus.test.scenarios.OpenShiftScenario;
-
-@OpenShiftScenario
-public class OpenShiftStrimziKafkaStreamIT extends StrimziKafkaStreamIT {
-}
