--- messaging/kafka-streams-reactive-messaging/src/test/java/io/quarkus/ts/openshift/messaging/kafka/OperatorOpenShiftAmqStreamsKafkaStreamIT.java
@@ -1,26 +1,0 @@
-package io.quarkus.ts.openshift.messaging.kafka;
-
-import org.junit.jupiter.api.Tag;
-
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.OpenShiftScenario;
-import io.quarkus.test.services.Operator;
-import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.test.services.operator.KafkaInstance;
-
-@Tag("operator-scenarios")
-@OpenShiftScenario
-public class OperatorOpenShiftAmqStreamsKafkaStreamIT extends BaseKafkaStreamIT {
-    @Operator(name = "amq-streams", source = "redhat-operators")
-    static KafkaInstance kafka = new KafkaInstance();
-
-    @QuarkusApplication
-    static RestService app = new RestService()
-            .withProperty("kafka.bootstrap.servers", kafka::getBootstrapUrl)
-            .withProperty("quarkus.kafka-streams.bootstrap-servers", kafka::getBootstrapUrl);
-
-    @Override
-    public RestService getApp() {
-        return app;
-    }
-}
