+++ security/keycloak/src/test/java/io/quarkus/ts/security/DevModeOidcSecurityIT.java
@@ -7,13 +7,13 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.junit.DisabledOnNativeImage;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.scenarios.annotations.DisabledOnNative;
 import io.quarkus.test.services.DevModeQuarkusApplication;
 
 @Tag("QUARKUS-1026")
 @QuarkusScenario
-@DisabledOnNativeImage
+@DisabledOnNative
 public class DevModeOidcSecurityIT {
 
     @DevModeQuarkusApplication
