+++ sql-db/sql-app/src/test/java/io/quarkus/ts/sqldb/sqlapp/DevModePostgresqlIT.java
@@ -3,13 +3,13 @@ package io.quarkus.ts.sqldb.sqlapp;
 import org.junit.jupiter.api.Tag;
 
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.junit.DisabledOnNativeImage;
 import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.scenarios.annotations.DisabledOnNative;
 import io.quarkus.test.services.DevModeQuarkusApplication;
 
 @Tag("QUARKUS-1026")
 @QuarkusScenario
-@DisabledOnNativeImage
+@DisabledOnNative
 public class DevModePostgresqlIT extends AbstractSqlDatabaseIT {
 
     @DevModeQuarkusApplication
